refactor: (#281) stub 파일 병합
