+++ dms-application/src/main/kotlin/team/aliens/dms/domain/studyroom/usecase/ManagerQueryStudyRoomsUseCase.kt
@@ -1,5 +1,6 @@
 package team.aliens.dms.domain.studyroom.usecase
 
+import java.util.UUID
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
 import team.aliens.dms.domain.school.validateSameSchool
 import team.aliens.dms.domain.studyroom.dto.ManagerQueryStudyRoomsResponse
@@ -9,7 +10,6 @@ import team.aliens.dms.domain.studyroom.spi.QueryStudyRoomPort
 import team.aliens.dms.domain.studyroom.spi.StudyRoomQueryUserPort
 import team.aliens.dms.domain.studyroom.spi.StudyRoomSecurityPort
 import team.aliens.dms.domain.user.exception.UserNotFoundException
-import java.util.UUID
 
 @ReadOnlyUseCase
 class ManagerQueryStudyRoomsUseCase(
