--- dms-application/src/test/kotlin/team/aliens/dms/domain/remain/stub/RemainOptionStub.kt
@@ -1,16 +1,0 @@
-package team.aliens.dms.domain.remain.stub
-
-import team.aliens.dms.domain.remain.model.RemainOption
-import java.util.UUID
-
-internal fun createRemainOptionStub(
-    id: UUID = UUID.randomUUID(),
-    schoolId: UUID = UUID.randomUUID(),
-    title: String = "title",
-    description: String = "description"
-) = RemainOption(
-    id = id,
-    schoolId = schoolId,
-    title = title,
-    description = description
-)
