--- dms-application/src/test/kotlin/team/aliens/dms/domain/school/stub/AvailableFeatureStub.kt
@@ -1,20 +1,0 @@
-package team.aliens.dms.domain.school.stub
-
-import team.aliens.dms.domain.school.model.AvailableFeature
-import java.util.UUID
-
-internal fun createAvailableFeatureStub(
-    schoolId: UUID = UUID.randomUUID(),
-    mealService: Boolean = true,
-    studyRoomService: Boolean = true,
-    remainService: Boolean = true,
-    pointService: Boolean = true,
-    noticeService: Boolean = true,
-) = AvailableFeature(
-    schoolId = schoolId,
-    mealService = mealService,
-    studyRoomService = studyRoomService,
-    remainService = remainService,
-    pointService = pointService,
-    noticeService = noticeService
-)
