+++ dms-application/src/test/kotlin/team/aliens/dms/domain/school/stub/SchoolStub.kt
@@ -1,8 +1,9 @@
 package team.aliens.dms.domain.school.stub
 
-import team.aliens.dms.domain.school.model.School
 import java.time.LocalDate
 import java.util.UUID
+import team.aliens.dms.domain.school.model.AvailableFeature
+import team.aliens.dms.domain.school.model.School
 
 internal fun createSchoolStub(
     id: UUID = UUID.randomUUID(),
@@ -22,4 +23,20 @@ internal fun createSchoolStub(
     address = address,
     contractStartedAt = contractStartedAt,
     contractEndedAt = contractEndedAt
+)
+
+internal fun createAvailableFeatureStub(
+    schoolId: UUID = UUID.randomUUID(),
+    mealService: Boolean = true,
+    studyRoomService: Boolean = true,
+    remainService: Boolean = true,
+    pointService: Boolean = true,
+    noticeService: Boolean = true,
+) = AvailableFeature(
+    schoolId = schoolId,
+    mealService = mealService,
+    studyRoomService = studyRoomService,
+    remainService = remainService,
+    pointService = pointService,
+    noticeService = noticeService
 )
